Migrate the db handling from `Gaia` to `Engine`
- Also move the scheduler cleanup in `Engine`
